+++ nemo/lightning/io/pl.py
@@ -23,6 +23,7 @@ from megatron.core.parallel_state import get_data_parallel_group
 from torch import nn
 from typing_extensions import Self, override
 
+from nemo.lightning.ckpt_utils import ckpt_to_dir
 from nemo.lightning.io.capture import IOProtocol
 from nemo.lightning.io.mixin import IOMixin
 
@@ -267,24 +268,6 @@ def _fix_tensors_device(ckpt: Dict) -> Dict:
         return t
 
     return dict_list_map_outplace(_fix_device, ckpt)
-
-
-def ckpt_to_dir(filepath: Union[str, Path]) -> Path:
-    """PTL considers checkpoints as .ckpt files.
-    This method removes the extension and returns a path
-    to be used as a directory for distributed checkpoints.
-    """
-    filepath = Path(filepath)
-    if not filepath.suffix == ".ckpt":
-        filepath = filepath.with_suffix(filepath.suffix + ".ckpt")
-
-    # adding this assert because we will later remove directories based on the return value of this method
-    assert filepath.suffix == ".ckpt", f"filepath: {filepath} must have .ckpt extension"
-
-    # create a new path whose name is the original filepath without the .ckpt extension
-    checkpoint_dir = filepath.with_name(filepath.stem)
-
-    return checkpoint_dir
 
 
 def is_distributed_ckpt(path) -> bool:
